% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_records.R
\name{create_records}
\alias{create_records}
\title{Create Records}
\usage{
create_records(object, data)
}
\arguments{
\item{object}{A string containing the object number. Ex: 'object_5'}

\item{data}{A dataframe with the data to upload. Column names should match field keys or labels.}
}
\value{
A status message. Ex: 'Success: (200) OK'
}
\description{
Create Records
}
\examples{
\dontrun{
# Example dataframe.
# Column names can be field numbers or labels.
data <- data.frame(Age = c(46,17,80),
                   Income = c(114, 63, 45.5),
                   HoursWk = c(60, 40, 42),
                   Language = c(0, 1, 0)
                   )

# This process may take some time to complete.
create_records('object_5', data)
}
}
