% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_records.R
\name{delete_records}
\alias{delete_records}
\title{Delete Records}
\usage{
delete_records(object, record_id)
}
\arguments{
\item{object}{A character string of the knack object. Ex: 'object_23'}

\item{record_id}{A character string of unique digits identifying the record. Ex: '61dc8e97403aa7116229c90a'}
}
\value{
A http status message
}
\description{
Delete Records
}
\examples{
\dontrun{
delete_records("object_23",
               "61dc8e97403aa7116229c90a")
}
}
