% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fields.R
\name{list_fields}
\alias{list_fields}
\title{List Fields of a Knack Object}
\usage{
list_fields(object, details = FALSE)
}
\arguments{
\item{object}{A string containing the knack object to retrieve fields from. Ex: 'object_23'}

\item{details}{A logical value saying whether to include all details.}
}
\value{
A data frame containing descriptions of fields in a Knack object
}
\description{
List Fields of a Knack Object
}
\examples{
\dontrun{
# List fields of a single object
list_fields("object_23")

# List fields of every object
objects <- list_objects()
sapply(objects$key, list_fields)
}

}
