% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_objects.R
\name{list_objects}
\alias{list_objects}
\title{List Knack Objects}
\usage{
list_objects()
}
\value{
A data frame listing the Knack Objects (i.e. Tables) available in the Knack Database
}
\description{
List Knack Objects
}
