% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_records.R
\name{n_records}
\alias{n_records}
\title{Determine the Number of Records in an Object}
\usage{
n_records(table_name)
}
\arguments{
\item{table_name}{A string containing the table name.}
}
\value{
An integer representing the number of records in the object.
}
\description{
Determine the Number of Records in an Object
}
