% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve Records From Knack}
\usage{
retrieve_data(object, include_raw = FALSE)
}
\arguments{
\item{object}{A string containing the knack object for which to retrieve data ex: 'object_23'}

\item{include_raw}{A logical value saying whether or not raw field values should be included. Default is no. Note that links and image files are only included in raw fields.}
}
\value{
A data frame (tibble)
}
\description{
Retrieve Records From Knack
}
\examples{
\dontrun{
retrieve_data('object_23','YOUR_API_ID','YOUR_API_KEY')
}
}
