% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_filtered_data.R
\name{retrieve_filtered_data}
\alias{retrieve_filtered_data}
\title{Send Filtered Data Requests to Knack}
\usage{
retrieve_filtered_data(
  object,
  api_id,
  api_key,
  include_raw = FALSE,
  filter_field = NA,
  match = NA,
  operator = NA,
  value = NA
)
}
\arguments{
\item{object}{A string containing the Knack object to pull data from. Ex: 'object_2'}

\item{api_id}{A string containing the Knack API ID.}

\item{api_key}{A string containing the Knack API key.}

\item{include_raw}{A logical stating whether raw fields should be included. Links are only included in raw fields.}

\item{filter_field}{A character vector of fields to filter. Can be one or more fields.}

\item{match}{A string containing either 'and' or 'or'. This will determine whether the records retrieved match all filters or at least one.}

\item{operator}{The character vector containing the operator for the filters. Options include:
contains, does not contain, is, is not, starts with, ends with, is blank, is not blank,
is during the current, is during the previous, is during the next, is before the previous,
is after the next, is before, is after, is today, is today or before, is today or after,
is before today, is after today, is before current time, is after current time.
Note that dates must be in the format 'YYYY-MM-DD'}

\item{value}{A character vector containing the values the operator uses to filter by.}
}
\value{
A data frame of filtered records.
}
\description{
Send Filtered Data Requests to Knack
}
\examples{
retrieve_filtered_data('object_2',
api_id = '61be439ed60d72001e68d749',
api_key = "57632271-982d-40ac-acf6-245b7f940dca",
filter_field = 'field_6',
match = 'or',
operator = c('starts with','ends with'),
value = c('h','i'))

retrieve_filtered_data("object_2",
api_id = "61be439ed60d72001e68d749",
api_key = "57632271-982d-40ac-acf6-245b7f940dca",
filter_field = "field_13",
match = "and",
operator = c("is after","if before"),
value = c("2021-12-2","2021-12-30"))
}
