% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_records.R
\name{retrieve_records}
\alias{retrieve_records}
\title{Retrieve Records}
\usage{
retrieve_records(table_name, n_rows = -1)
}
\arguments{
\item{table_name}{The name of the table to retrieve records from.}

\item{n_rows}{The number of rows to retrieve. The default is all rows.}
}
\value{
A data frame of record values.
}
\description{
Retrieve Records
}
