% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Set API Authentication Credentials}
\usage{
set_credentials(api_id, api_key)
}
\arguments{
\item{api_id}{A string containing the Knack API ID}

\item{api_key}{A string containing the Knack API Key}
}
\value{
No value is returned. API ID and API Key are set in the R session.
}
\description{
Set API Authentication Credentials
}
