% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_view_based_credentials.R
\name{set_view_based_credentials}
\alias{set_view_based_credentials}
\title{Set View-Based Knack Credentials}
\usage{
set_view_based_credentials(email, password, api_id)
}
\arguments{
\item{email}{A string containing the login email address.}

\item{password}{A string containing the login password.}

\item{api_id}{A string containing the Knack API ID.}
}
\value{

}
\description{
Set View-Based Knack Credentials
}
\examples{
\dontrun{
set_view_based_credentials('example@example.com',
                           'verysecretpassword',
                           '60a2a941bc3064001b208e21')

}
}
