% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_records.R
\name{update_records}
\alias{update_records}
\title{Update Records}
\usage{
update_records(object, record_id, data)
}
\arguments{
\item{object}{A character string containing the object number. Ex: 'object_23'}

\item{record_id}{A character string of the record id. Ex: '61d87cf997ff7a33aaf44b6b'}

\item{data}{A data frame of the values to change. Column names should be the field ID's.}
}
\value{
A http status message.
}
\description{
Update Records
}
\examples{
\dontrun{
update_records("object_23",
               "61d87cf997ff7a33aaf44b6b",
               data.frame(field_170 = 2, field_168 = "New Data"))
}
}
