% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fields.R
\name{list_fields}
\alias{list_fields}
\title{List Fields and Data Types of a Knack Object}
\usage{
list_fields(table_name)
}
\arguments{
\item{table_name}{The name of the table to query.}
}
\value{
A data frame containing descriptions of fields.
}
\description{
List Fields and Data Types of a Knack Object
}
