% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_records.R
\name{n_records}
\alias{n_records}
\title{Determine the Number of Records in an Object}
\usage{
n_records(object)
}
\arguments{
\item{object}{A string containing the object number. Ex: 'object_2'}
}
\value{
An integer representing the number of records in the object.
}
\description{
Determine the Number of Records in an Object
}
\examples{
\dontrun{
n_records('object_2')
}
}
