% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_records.R
\name{retrieve_records}
\alias{retrieve_records}
\title{Retrieve Records From Knack}
\usage{
retrieve_records(
  object,
  filter_field = "",
  value = "",
  operator = "is",
  match = "and",
  include_raw = FALSE,
  limit = 1000
)
}
\arguments{
\item{object}{A string containing the Knack object to pull data from. Ex: 'object_2'}

\item{filter_field}{A character vector of fields to filter. Can be one or more fields.}

\item{value}{A character vector containing the values the operator uses to filter by.}

\item{operator}{The character vector containing the operator for the filters. Options include:
contains, does not contain, is, is not, starts with, ends with, is blank, is not blank,
is during the current, is during the previous, is during the next, is before the previous,
is after the next, is before, is after, is today, is today or before, is today or after,
is before today, is after today, is before current time, is after current time, is lower than, is higher than.
Note that dates must be in the format 'YYYY-MM-DD'}

\item{match}{A string containing either 'and' or 'or'. This will determine whether the records retrieved match all filters or at least one.}

\item{include_raw}{A logical stating whether raw fields should be included. Links are only included in raw fields.}

\item{limit}{The number of records to include per page. The default is 1000.}
}
\value{
A data frame of filtered records.
}
\description{
Retrieve Records From Knack
}
\examples{
\dontrun{
# Retrieve all records from an object
retrieve_records("object_23")

# You can quick retrieve records according to a single condition
retrieve_records("object_23", "job_no", "0121-023")

# Retrieve records within a date range
# Note that dates must be in 'YYYY-MM-DD' format
retrieve_records("object_23",
                  "date",
                  c('2021-12-01','2021-12-30'),
                  c("is after","is before"))


# Retrieve records according to multiple field conditions
retrieve_records("object_23",
                  c("day_of_week","hours_worked", "account"),
                  c("Monday",5, "Marcus")
                  c("is","is greater than", "contains"))

}

}
